rename `StreamExecFirstLastRow` to `StreamExecDeduplicate`, and do not convert sql like `SELECT a FROM MyTable ORDER BY proctime LIMIT 1` to `StreamExecDeduplicate`
