+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/plan/rules/FlinkStreamRuleSets.scala
@@ -231,8 +231,7 @@ object FlinkStreamRuleSets {
     StreamExecSortLimitRule.INSTANCE,
     StreamExecRankRule.INSTANCE,
     StreamExecTemporalSortRule.INSTANCE,
-    StreamExecFirstLastRowRule.SORT_INSTANCE,
-    StreamExecFirstLastRowRule.RANK_INSTANCE,
+    StreamExecDeduplicateRule.RANK_INSTANCE,
     StreamExecCorrelateRule.INSTANCE,
     StreamExecSinkRule.INSTANCE
   )
